ask_hr_agent/workday_tools/requirements.txt
docxtpl==0.16.7
fastapi==0.123.10
google-adk==1.21.0
httpx==0.28.1
Jinja2==3.1.6
python-docx==1.1.2
requests==2.32.5
selenium==4.39.0
uvicorn==0.40.0
